+++ app/com/atlassian/connect/play/java/auth/jwt/PlayRequestWrapper.java
@@ -2,6 +2,7 @@ package com.atlassian.connect.play.java.auth.jwt;
 
 import com.atlassian.jwt.CanonicalHttpRequest;
 import com.atlassian.jwt.core.http.HttpRequestWrapper;
+import com.google.common.collect.ImmutableList;
 import org.apache.commons.lang.StringUtils;
 import play.mvc.Http;
 
@@ -28,7 +29,18 @@ public class PlayRequestWrapper implements HttpRequestWrapper {
 
     @Override
     public Iterable<String> getHeaderValues(String headerName) {
-        return Arrays.asList(request.headers().get(headerName));
+        // Header retrieval must be case insensitive
+        // (http://www.w3.org/Protocols/HTTP/1.1/rfc2616bis/draft-lafon-rfc2616bis-02.html#message.headers)
+        // Note the following code was copied and adapted from RequestHeader.getHeader()
+        String[] headerValues = null;
+        final Map<String, String[]> headers = request.headers();
+        for(String h: headers.keySet()) {
+            if(headerName.toLowerCase().equals(h.toLowerCase())) {
+                headerValues = headers.get(h);
+                break;
+            }
+        }
+        return headerValues != null ? Arrays.asList(headerValues) : ImmutableList.<String>of();
     }
 
     @Override