+++ changelog.md
@@ -1,5 +1,21 @@
 # Changelog - Play Java Module for Atlassian Connect
 
+## 0.8.2
+
+* Bug Fix: Removed `@Required` from _sharedSecret_ and _publicKey_ as they are not required for auth type *none*
+
+## 0.8.1
+
+* Fixed header lookup to be case insensitive. Caused issues with Authorization header in some cases
+* Fixed NPE in PlayRequestWrapper
+
+## 0.8.0
+
+* **BREAKING**
+
+    * During JWT authentication, the full base URL (including any path) is excluded from the qsh calculation
+    * URL encoding of canonical path separator (&)
+
 ## 0.7.1
 
 * changed to use new AUI CDN